Points can be negative

+++ src/FPL/Reports/Player.hs
@@ -61,7 +61,7 @@ playerExpectedPoints TeamStats {..} FixtureExpected {..} PlayerStats {..} =
       psGoals * pointsForGoal psPosition
         + psAssists * pointsForAssist
     totalPtsFromCS = psCleanSheets * pointsForCS psPosition
-    totalOtherPts = psPoints - (totalPtsFromGoals + totalPtsFromCS)
+    totalOtherPts = psPoints - fromIntegral (totalPtsFromGoals + totalPtsFromCS)
 
 playersOverview :: PlayersData -> [[T.Text]]
 playersOverview = \PlayersData {..} ->
